UBUNTU: SAUCE: apparmor: Fix: insert race between label_update and label_merge
BugLink: http://bugs.launchpad.net/bugs/1448912

it is possible for a label_merge to be racing label_update resulting
in the AA_BUG(__aa_label_remove_and_insert()) to trip. Instead check
if a version of label has already been inserted and gracefully fail
if label_update lost the race.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>