UBUNTU: SAUCE: apparmor: Fix: convert replacedby update to be protected by the labelset lock
BugLink: http://bugs.launchpad.net/bugs/1448912

replacedby updates must be able to occur when in an rcu critical sections,
and when spin locks are held. As such it can not use a mutex lock to
protect its critical section. Since replacedby updates are accompanied by
labelset insertion and removals use the labelset write lock to protect
the update critical section.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>